moved run code from bin to lib
